:zap: :bug: fixture on PUT route when modifying a sauce

+++ controllers/sauce.js
@@ -53,16 +53,25 @@ exports.modifySauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then(oldSauce => {
       const oldImg = oldSauce.imageUrl.split('/img/')[1];
-      const sauceObject = { 
-        ...JSON.parse(req.body.sauce), 
-        imageUrl: `${req.protocol}://${req.get('host')}/img/${req.file.filename}`
-        } // : { ...req.body };
-        // to remove unnecesary images
+      const sauceObject = req.file ? 
+        { ...JSON.parse(req.body.sauce), 
+          imageUrl: `${req.protocol}://${req.get('host')}/img/${req.file.filename}` 
+        } : { 
+          ...req.body 
+        };
+      if (req.file) {
+        // to remove unnecesary old images
         fs.unlink('img/' + oldImg, () => {
           Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
             .then(() => res.status(200).json({ message: 'The sauce was updated' }))
             .catch(error => res.status(400).json({ error }));
         });
+      } else {
+        // Only update sauce fields without touching the image 
+        Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+          .then(() => res.status(200).json({ message: 'The sauce was updated' }))
+          .catch(error => res.status(400).json({ error }));
+      }
     })
     .catch(error => res.status(500).json({ error }));
 };
